+++ xdas/core/dataarray.py
@@ -12,7 +12,7 @@ import numpy as np
 import xarray as xr
 from numpy.lib.mixins import NDArrayOperatorsMixin
 
-from ..dask import from_dict, to_dict
+from ..dask.core import from_dict, to_dict
 from ..virtual import VirtualArray, VirtualSource, _to_human
 from .coordinates import Coordinate, Coordinates, get_sampling_interval
